feat: update application and static generator
