+++ Sources/App/main.swift
@@ -1,0 +1,4 @@
+import BlogPublish
+
+try App()
+    .run()
