+++ Sources/BlogPublish/BlogPublish.swift
@@ -1,7 +1,13 @@
+//
+//  File.swift
+//  
+//
+//  Created by Amine Bensalah on 09/10/2021.
+//
+
 import Foundation
 import Publish
 import Plot
-import SplashPublishPlugin
 
 enum Constants {
     static let website = "aminebensalah.com"
@@ -48,26 +54,3 @@ struct BlogPublish: Website {
     var language: Language { .english }
     var imagePath: Path? { nil }
 }
-
-// This will generate your website using the built-in Foundation theme:
-try BlogPublish().publish(using: [
-    .copyResources(),
-    .installPlugin(.splash(withClassPrefix: "")),
-    .installPlugin(.ensureAllItemsAreTagged),
-    .addMarkdownFiles(),
-    .addDefaultSectionTitles(),
-    .sortItems(by: \.date, order: .descending),
-    .generateHTML(withTheme: .basic),
-    .unwrap(RSSFeedConfiguration.default, { config in
-        let matcher = { (item: Item<BlogPublish>) -> Bool in
-            item.metadata.published
-        }
-        let predicate = Predicate<Item<BlogPublish>>(matcher: matcher)
-        return .generateRSSFeed(including: [.posts, .tips],
-                                itemPredicate: predicate,
-                                config: config,
-                                date: Date())
-    }),
-    .createCNAME(website: Constants.website),
-    .generateSiteMap()
-])
