+++ Tests/BlogPublishTests/BlogPublishTests.swift
@@ -1,4 +1,5 @@
 import XCTest
+@testable import BlogPublish
 
 class BlogPublishTests: XCTestCase {
